Update configure-fastly.md
Magento setup:store-config: set command  has 2 options for setting the base URL --base-url="http://mcstaging.your-domain.com/ for HTTP and --base-url-secure for HTTPS. It is best practice to change both of these values, even with "Force TLS" options enabled. The current suggestion is incorrect, because we are setting the base unsecure URL to https://mcstaging.your-domain.com/" while leaving SECURE (HTTPS) unchanged. The customers launches on the SECURE URL and with FORCE TLS enabled, all traffic would be served over HTTPS so the suggestion we mnake isnt used most of the times. We would need to set these values accoidngingly with each respective values and enable FORCE TLS https://devdocs.magento.com/cloud/cdn/configure-fastly-customize-cache.html#force-tls. it would be great if we could make this adjustment, as what we are showing currently is incorrect.

For your reference this is the database paths for those options:

| config_id | scope   | scope_id | path                  | value                                    | updated_at          |
+-----------+---------+----------+-----------------------+------------------------------------------+---------------------+
|         1 | default |        0 | web/unsecure/base_url | http://mcstaging.your-domain.com/ | 2021-06-17 15:35:42 |
|         2 | default |        0 | web/secure/base_url   | https://mcstaging.your-domain.com/ | 2021-06-17 15:42:08 |
+-----------+---------+----------+-----------------------+------------------------------------------+---------------------+